refactor: move routes logic to routes file
