+++ src/App.tsx
@@ -1,21 +1,15 @@
 import { useCallback, useEffect } from 'react'
 import 'antd/dist/antd.min.css'
 import './App.scss'
-import TimelinePage from 'src/pages/TimelinePage'
 import { ConfigProvider, Layout } from 'antd'
 import 'leaflet/dist/leaflet.css'
-import { TEXT_KEYS } from 'src/resources/texts'
 import styled from 'styled-components'
 import heIL from 'antd/es/locale/he_IL'
-import { BrowserRouter as Router, Navigate, Route, Routes, useSearchParams } from 'react-router-dom'
-import GapsPage from './pages/GapsPage'
+import { BrowserRouter as Router, useSearchParams } from 'react-router-dom'
 import { PageSearchState, SearchContext } from './model/pageState'
 import moment from 'moment'
-import DashboardPage from './pages/dashboard/DashboardPage'
 import { useSessionStorage } from 'usehooks-ts'
 import SideBar from './pages/components/header/sidebar/SideBar'
-import RealtimeMapPage from './pages/RealtimeMapPage'
-import SingleLineMapPage from './pages/SingleLineMapPage'
 import { useLocation } from 'react-router-dom'
 import ReactGA from 'react-ga4'
 import { CacheProvider } from '@emotion/react'
@@ -26,20 +20,8 @@ import { heIL as heILmui } from '@mui/x-date-pickers/locales'
 import { ThemeProvider, createTheme } from '@mui/material'
 import { AdapterMoment } from '@mui/x-date-pickers/AdapterMoment'
 import { LocalizationProvider } from '@mui/x-date-pickers'
-import About from './pages/About'
-import GapsPatternsPage from './pages/GapsPatternsPage'
-import {
-  RadarChartOutlined,
-  BellOutlined,
-  DollarOutlined,
-  HeatMapOutlined,
-  LaptopOutlined,
-  FieldTimeOutlined,
-  BugOutlined,
-  BarChartOutlined,
-  LineChartOutlined,
-} from '@ant-design/icons'
-import { MenuPage } from './pages/components/header/menu/Menu'
+
+import RoutesList, { PAGES } from './routes'
 const { Content } = Layout
 
 const StyledLayout = styled(Layout)`
@@ -54,57 +36,6 @@ const StyledBody = styled.div`
   padding: 24px;
   min-height: 360px;
 `
-
-const PAGES = [
-  {
-    label: TEXT_KEYS.dashboard_page_title,
-    key: '/dashboard',
-    icon: LaptopOutlined,
-  },
-  {
-    label: TEXT_KEYS.timeline_page_title,
-    key: '/timeline',
-    searchParamsRequired: true,
-    icon: FieldTimeOutlined,
-  },
-  {
-    label: TEXT_KEYS.gaps_page_title,
-    key: '/gaps',
-    searchParamsRequired: true,
-    icon: BarChartOutlined,
-  },
-  {
-    label: TEXT_KEYS.gaps_patterns_page_title,
-    key: '/gaps_patterns',
-    icon: LineChartOutlined,
-  },
-  {
-    label: TEXT_KEYS.realtime_map_page_title,
-    key: '/map',
-    icon: HeatMapOutlined,
-  },
-  {
-    label: TEXT_KEYS.singleline_map_page_title,
-    key: '/single-line-map',
-    searchParamsRequired: true,
-    icon: RadarChartOutlined,
-  },
-  {
-    label: TEXT_KEYS.about_title,
-    key: '/about',
-    icon: BellOutlined,
-  },
-  {
-    label: TEXT_KEYS.report_a_bug_title,
-    key: 'https://github.com/hasadna/open-bus-map-search/issues',
-    icon: BugOutlined,
-  },
-  {
-    label: TEXT_KEYS.donate_title,
-    key: 'https://www.jgive.com/new/he/ils/donation-targets/3268#donation-modal',
-    icon: DollarOutlined,
-  },
-] as MenuPage[]
 
 const theme = createTheme(
   {
@@ -169,8 +100,6 @@ const App = () => {
     })
   }, [])
 
-  const RedirectToDashboard = () => <Navigate to={PAGES[0].key} replace />
-
   return (
     <SearchContext.Provider value={{ search, setSearch: safeSetSearch }}>
       <CacheProvider value={cacheRtl}>
@@ -182,16 +111,7 @@ const App = () => {
                 <Layout>
                   <StyledContent>
                     <StyledBody>
-                      <Routes>
-                        <Route path={PAGES[0].key} element={<DashboardPage />} />
-                        <Route path={PAGES[1].key} element={<TimelinePage />} />
-                        <Route path={PAGES[2].key} element={<GapsPage />} />
-                        <Route path={PAGES[3].key} element={<GapsPatternsPage />} />
-                        <Route path={PAGES[4].key} element={<RealtimeMapPage />} />
-                        <Route path={PAGES[5].key} element={<SingleLineMapPage />} />
-                        <Route path={PAGES[6].key} element={<About />} />
-                        <Route path="*" element={<RedirectToDashboard />} />
-                      </Routes>
+                      <RoutesList />
                     </StyledBody>
                   </StyledContent>
                 </Layout>
